LayoutTests/media/video-src-set-expected.txt

Test that setting src attribute triggers load
EVENT(begin) TEST(relativeURL(video.currentSrc) == 'content/test.mp4') OK
END OF TEST
